Default DPS removed (it now falls back to a sensible number of dps) HexToRGB takes an opacity input parseInt replaced with parseFloat for barcharts KPI chart now formats metrics Pie chart values / tooltips / labels formatted Stacked area chart date formats updated Test float data added to models

+++ src/components/vanilla/charts/BarChart/index.tsx
@@ -82,7 +82,7 @@ function chartData(props: Props): ChartData<'bar'> {
         minBarLength: 0,
         borderRadius: 6,
         label: metric.title,
-        data: results?.data?.map((d) => parseInt(d[metric.name])) || [],
+        data: results?.data?.map((d) => parseFloat(d[metric.name])) || [],
         backgroundColor: COLORS[i % COLORS.length]
       })) || []
   };
